+++ src/OpenTelemetry.Extensions/Trace/RateLimitingSampler.cs
@@ -28,6 +28,15 @@ public class RateLimitingSampler : Sampler
     /// <param name="maxTracesPerSecond">The maximum number of traces that will be emitted each second.</param>
     public RateLimitingSampler(int maxTracesPerSecond)
     {
+#if NET
+        ArgumentOutOfRangeException.ThrowIfLessThanOrEqual(maxTracesPerSecond, 0, nameof(maxTracesPerSecond));
+#else
+        if (maxTracesPerSecond <= 0)
+        {
+            throw new ArgumentOutOfRangeException(nameof(maxTracesPerSecond), maxTracesPerSecond, "Value must be greater than zero.");
+        }
+#endif
+
         var maxBalance = maxTracesPerSecond < 1.0 ? 1.0 : maxTracesPerSecond;
         this.rateLimiter = new RateLimiter(maxTracesPerSecond, maxBalance);
         var attributes = new Dictionary<string, object>()