+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -5,6 +5,13 @@
 * Updated OpenTelemetry core component version(s) to `1.15.2`.
   ([#4080](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4080))
 
+* Fix `IndexOutOfRangeException` when parsing SQL statements.
+  ([#4139](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4139))
+
+* Add instrumentation schema URL to traces when using either the old or new
+  database Semantic Conventions, but not when both are used together.
+  ([#4078](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4078))
+
 ## 1.15.0-beta.1
 
 Released 2026-Jan-21